requirements.txt
beautifulsoup4
pytelegrambotapi
python-dotenv
schedule
moviepy
pymongo
google-generativeai
